integrations/ctrader.py
import requests
import websocket

class CTraderIntegration:
    def __init__(self, client_id, client_secret, access_token=None, refresh_token=None):
        self.client_id = client_id
        self.client_secret = client_secret
        self.access_token = access_token
        self.refresh_token = refresh_token
        self.connected = False

    def connect(self):
        # TODO: Implement OAuth2 authentication and WebSocket connection
        # See: https://connect.spotware.com/apps/open_api
        pass

    def get_account_info(self):
        # TODO: Implement account info retrieval via REST API
        pass

    def get_historical_data(self, symbol, timeframe, start, end):
        # TODO: Implement historical data fetch via REST API
        pass

    def send_order(self, symbol, volume, order_type, price=None, sl=None, tp=None, comment="AI Trade"):
        # TODO: Implement order sending via WebSocket API
        pass 