requirements.txt
gs-quant>=1.0.0
yfinance
ta
textblob
finviz
scikit-learn
xgboost
lightgbm
tensorflow
torch
streamlit

# Core
pandas
numpy
requests
aiohttp

# ML & Analysis
scikit-learn
vaderSentiment
textblob
nltk

# Word Embeddings
gensim
emoji

# Twitter
tweepy

# Environment Variables
python-dotenv

# Gemini / Google Generative AI
google-generativeai

# Backtesting
# Install qs-quant directly from GitHub
backtrader
ta


# Optional (if needed for plotting or logs)
matplotlib
loguru

MetaTrader5

websocket-client
